Added updateRecurring trigger
