+++ backend/lambda/updateRecurring/requirements.txt
@@ -1,0 +1,11 @@
+boto3==1.17.7
+botocore==1.20.7
+certifi==2020.12.5
+chardet==4.0.0
+idna==2.10
+jmespath==0.10.0
+python-dateutil==2.8.1
+requests==2.25.1
+s3transfer==0.3.4
+six==1.15.0
+urllib3==1.26.3
